Add patch tabs_privileged_callback.patch and update docs for it
In previous menu tweak(dcb5f6d209) for RT 5, we used a couple of changes
not available in RT 4.4.1 and prior versions, which is what the patch
for.

As RT 4.0 has already reached EOL for a while, we don't need to worry
about it and can simply bump required RT version.