Fix missing Calendar.html Format when accessing from MyCalendar
There was a situation where the Download Spreadsheet link was not
working due to the Format attribute not being set.  This comes from
the default "calendar" search and was only setting the TempFormat
variable. Fixed that by setting the Format variable with the TempoFormat
if not set yet.

Also moved the Query String builder, which is used by the Filter On Status
form and the link to change month and was also missing the Format
attribute.