Fix display of multiple days events
The loop on the code that was checking if an event had multiple days had
a minor bug because it was adding one day to each loop but using
unixtime to check if the event had the next day ahead. That was causing
the last day of the event to be disconnected from the rest if it was
starting at a time lower than the start time of the event.