Fix crash when changing months using the select boxes
After changing a month using the nav links, the query string
used to pre-populate the action of the form and other elements
already had a month and year. Since the select boxes for month
and year had the same name, both variables were considered arrays
and failed to be parsed.