Allow subject content to overflow on multi-day entries
On a single day entry, the text of the anchor is cut off
without the max-width. With the max-width, it is also cut
off on multi-day events even though there is room in the
event bar. Removing max-width allows the text to overflow
and show in the available space on the next day.